refactor: extract user details fetching logic to fetchUserDetails.js for reusability
